refactor: #80 Refactor Auth Interceptor
