+++ src/main/java/org/greenbyme/angelhack/auth/AuthInterceptor.java
@@ -1,6 +1,8 @@
-package org.greenbyme.angelhack.auth.jwt;
+package org.greenbyme.angelhack.auth;
 
 import lombok.extern.slf4j.Slf4j;
+import org.greenbyme.angelhack.exception.ErrorCode;
+import org.greenbyme.angelhack.exception.UserException;
 import org.greenbyme.angelhack.util.JwtTokenProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -18,11 +20,11 @@ public class AuthInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
-            throws Exception {
+            throws UserException {
         final String token = request.getHeader("jwt");
-        log.info("preHandle:" + token);
-        if (jwtTokenProvider.validateToken(token)) {
-            throw new IllegalAccessException("권한이 없습니다");
+        log.info("preHandle: " + token);
+        if (!jwtTokenProvider.validateToken(token)) {
+            throw new UserException("권한이 없습니다", ErrorCode.INVALID_ACCESS);
         }
         request.setAttribute("userID", jwtTokenProvider.getId(token));
         return true;
