+++ src/main/java/org/greenbyme/angelhack/config/WebMvcConfig.java
@@ -1,21 +1,24 @@
 package org.greenbyme.angelhack.config;
 
-import org.greenbyme.angelhack.auth.jwt.AuthInterceptor;
+import org.greenbyme.angelhack.auth.AuthInterceptor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import java.util.Arrays;
-import java.util.Collections;
-
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
 
     private static final String[] EXCLUDE_PATHS = {
             "/api/users/signin",
-            "/api/users"};
+            "/api/users/signup",
+            "/api/users/images/{fileName}",
+            "/api/missions/**",
+            "/api/categorys/**",
+            "/api/post/images/**",
+            "/api/post/missions/**"
+            };
 
     private final long MAX_AGE_SECS = 3600;
